@@ -1,12 +1,17 @@
-import React from 'react';
-import ReactDOM from 'react-dom';
-import App from './App';
-import Store from "./Redux/Store"
-import {Provider} from "react-redux"
+import React from "react";
+import ReactDOM from "react-dom/client";
+import './index.css'
+import App from "./App";
+import { Provider } from "react-redux";
+// import reportWebVitals from "./reportWebVitals";
+import Store from './Redux/Store'
 
-ReactDOM.render(
+const root = ReactDOM.createRoot(document.getElementById("root"));
+root.render(
   <Provider store={Store}>
-    <App/>
-  </Provider>,
-  document.getElementById('root')
+  <React.StrictMode>
+    <App />
+  </React.StrictMode>
+  </Provider>
 );
+
